Ref issue #10 A script to bootstrap a SOLVCON parallel cluster on GCE.
This script is the basic component to bootstrap GCE instances to be a
very simple Beowulf cluster. Besides, this cluster could be used for
SOLVCON parallel computing with this bootstrap script. This commit is
the 1st step of the whole story of the issue #10. Wrappers to use this
bootstrap script will follow up in the later commits.