Ref issue 12 add script remark.
This commit is according to the discussion of PR 11 to add the PR
description to be the script remark. This commit is related to issue 10
of project 1 as well. I also tweak the format and logic to show the
usage description to make this script be more readable and maintainable.